requirements.txt
asgiref==3.8.1
Django==5.1
django-log-viewer==1.1.7
gunicorn==23.0.0
mysqlclient==2.2.4
python-dotenv==1.0.1
RouterOS-api==0.18.1
sqlparse==0.5.1
tzdata==2024.1
